# CANN: Optimize MUL_MAT_ID